+++ features/step_definitions/cli_steps.rb
@@ -16,8 +16,8 @@ When /^I have a valid project on localeapp\.com but an incorrect api key "([^"]*
 end
 
 When /^I have a translations on localeapp\.com for the project with api key "([^"]*)"$/ do |api_key|
-  uri = "https://api.localeapp.com/v1/projects/#{api_key}/translations.yml"
-  body = valid_translation_data.to_yaml
+  uri = "https://api.localeapp.com/v1/projects/#{api_key}/translations/all.yml"
+  body = valid_export_data.to_yaml
   add_fake_web_uri(:get, uri, ['200', 'OK'], body)
 end
 
@@ -25,6 +25,13 @@ When /^new translations for the api key "([^"]*)" since "([^"]*)" with time "([^
   uri = "https://api.localeapp.com/v1/projects/#{api_key}/translations.yml?updated_at=#{update_time}"
   body = valid_translation_data.to_yaml
   add_fake_web_uri(:get, uri, ['200', 'OK'], body, 'date' => Time.at(new_time.to_i).httpdate)
+end
+
+
+When /^I have a valid project on localeapp\.com with api key "([^"]*)" and the translation key "([^"]*)"/ do |api_key, key_name|
+  uri = "https://api.localeapp.com/v1/projects/#{api_key}/translations/#{key_name.gsub(/\./, '%2E')}"
+  add_fake_web_uri(:delete, uri, ['200', 'OK'], '')
+  add_fake_web_uri(:post, uri + '/rename', ['200', 'OK'], '')
 end
 
 When /^an initializer file$/ do