+++ lib/localeapp.rb
@@ -35,11 +35,15 @@ require 'localeapp/key_checker'
 require 'localeapp/missing_translations'
 require 'localeapp/default_value_handler'
 
+require 'localeapp/cli/command'
 require 'localeapp/cli/install'
 require 'localeapp/cli/pull'
 require 'localeapp/cli/push'
 require 'localeapp/cli/update'
 require 'localeapp/cli/add'
+require 'localeapp/cli/remove'
+require 'localeapp/cli/rename'
+require 'localeapp/cli/daemon'
 
 # AUDIT: Will this work on ruby 1.9.x
 $KCODE="UTF8" if RUBY_VERSION < '1.9'
@@ -73,6 +77,10 @@ module Localeapp
       logger.info LOG_PREFIX + message if logger
     end
 
+    def log_with_time(message)
+      log [Time.now.to_i, message].join(' - ')
+    end
+
     def debug(message)
       logger.debug(LOG_PREFIX + message) if logger
     end
@@ -88,27 +96,22 @@ module Localeapp
     # end
     def configure
       self.configuration ||= Configuration.new
-      yield(configuration)
+      yield(configuration) if block_given?
       self.sender  = Sender.new
       self.poller  = Poller.new
       self.updater = Updater.new
       @missing_translations = MissingTranslations.new
     end
 
-    # requires the Localeapp configuration
-    def initialize_config(file_path=nil)
-      file_paths = [ File.join(Dir.pwd, '.localeapp', 'config.rb'),
-                     File.join(Dir.pwd, 'config', 'initializers', 'localeapp.rb') ]
-      file_paths << file_path if file_path
-      file_paths.each do |path|
-        next unless File.exists? path
-        begin
-          require path
-          return true
-        rescue
-        end
-      end
-      false
+    def has_config_file?
+      default_config_file_paths.any? { |path| File.exists?(path) }
+    end
+
+    def default_config_file_paths
+      [
+        File.join(Dir.pwd, '.localeapp', 'config.rb'),
+        File.join(Dir.pwd, 'config', 'initializers', 'localeapp.rb')
+      ]
     end
 
     def load_yaml(contents)
@@ -125,8 +128,16 @@ module Localeapp
 
     private
 
+    def private_null_type(results)
+      return true if results.is_a?(YAML::PrivateType) && results.type_id == 'null'
+      if RUBY_PLATFORM == 'java'
+        return true if results.is_a?(YAML::Yecht::PrivateType) && results.type_id == 'null'
+      end
+      false
+    end
+
     def normalize_results(results)
-      if results.is_a?(YAML::PrivateType) && results.type_id == 'null'
+      if private_null_type(results)
         nil
       elsif results.is_a?(Array)
         results.each_with_index do |value, i|