+++ lib/localeapp/cli/pull.rb
@@ -1,18 +1,14 @@
 module Localeapp
   module CLI
-    class Pull
+    class Pull < Command
       include ::Localeapp::ApiCall
-
-      def initialize(output = $stdout)
-        @output = output
-      end
 
       def execute
         @output.puts "Localeapp Pull"
         @output.puts ""
 
         @output.puts "Fetching translations:"
-        api_call :translations,
+        api_call :export,
           :success => :update_backend,
           :failure => :report_failure,
           :max_connection_attempts => 3
@@ -21,7 +17,7 @@ module Localeapp
       def update_backend(response)
         @output.puts "Success!"
         @output.puts "Updating backend:"
-        Localeapp.updater.update(Localeapp.load_yaml(response))
+        Localeapp.updater.dump(Localeapp.load_yaml(response))
         @output.puts "Success!"
         Localeapp.poller.write_synchronization_data!(Time.now.to_i, Time.now.to_i)
       end