+++ lib/localeapp/default_value_handler.rb
@@ -3,15 +3,12 @@ module I18n::Backend::Base
 
   def default(locale, object, subject, options = {})
     result = default_without_handler(locale, object, subject, options)
-    return result if ::Localeapp.configuration.sending_disabled?
-
-    if result
-      sender = Localeapp::Sender.new
-
-      # Make the default value a complete translation
-      sender.post_translation(locale, object, options, result)
+    case subject # case is what i18n gem uses here so doing the same
+    when Array
+      # Do nothing, we only send missing translations with text defaults
+    else
+      Localeapp.missing_translations.add(locale, object, subject, options)
     end
-
     return result
   end
 end