+++ lib/localeapp/routes.rb
@@ -1,3 +1,5 @@
+require 'rack/utils'
+
 module Localeapp
   module Routes
     VERSION = 'v1'
@@ -24,6 +26,35 @@ module Localeapp
 
     def create_translation_endpoint(options = {})
       [:post, translations_url(options)]
+    end
+
+    def export_endpoint(options = {})
+      [:get, export_url(options)]
+    end
+
+    def remove_endpoint(options = {})
+      [:delete, remove_url(options)]
+    end
+
+    def remove_url(options = {})
+      url = http_scheme.build(base_options.merge(:path => remove_path(options[:key], options[:format])))
+      url.to_s
+    end
+
+    def rename_endpoint(options = {})
+      [:post, rename_url(options)]
+    end
+
+    def rename_url(options = {})
+      url = http_scheme.build(base_options.merge(:path => rename_path(options[:current_name], options[:format])))
+      url.to_s
+    end
+
+    def export_url(options = {})
+      options[:format] ||= 'yml'
+      url = http_scheme.build(base_options.merge(:path => export_path(options[:format])))
+      url.query = options[:query].map { |k,v| "#{k}=#{v}" }.join('&') if options[:query]
+      url.to_s
     end
 
     def missing_translations_endpoint(options = {})
@@ -70,6 +101,30 @@ module Localeapp
 
     def translations_path(format = nil)
       path = project_path << '/translations'
+      path << ".#{format}" if format
+      path
+    end
+
+    def remove_path(key, format = nil)
+      raise "remove_path requires a key" if key.nil?
+      path = translations_path << "/#{escape_key(key)}"
+      path << ".#{format}" if format
+      path
+    end
+
+    def rename_path(current_name, format = nil)
+      raise "rename_path requires current name" if current_name.nil?
+      path = translations_path << "/#{escape_key(current_name)}" << '/rename'
+      path << ".#{format}" if format
+      path
+    end
+
+    def escape_key(key)
+      Rack::Utils.escape(key).gsub(/\./, '%2E')
+    end
+
+    def export_path(format = nil)
+      path = project_path << '/translations/all'
       path << ".#{format}" if format
       path
     end