+++ spec/localeapp/api_caller_spec.rb
@@ -36,6 +36,22 @@ describe Localeapp::ApiCaller, "#call(object)" do
     @api_caller.call(self)
   end
 
+  if "".respond_to?(:force_encoding)
+    def success_check(response)
+      response.encoding.should == Encoding.find('UTF-8')
+    end
+
+    it "sets the response encoding based on the response charset" do
+      response = "string"
+      response.stub!(:code).and_return(200)
+      response.force_encoding('US-ASCII')
+      response.stub_chain(:net_http_res, :type_params).and_return('charset' => 'utf-8')
+      RestClient::Request.stub(:execute).and_return(response)
+      @api_caller.options[:success] = :success_check
+      @api_caller.call(self)
+    end
+  end
+
   context "Proxy" do
     before do
       RestClient::Request.stub!(:execute).and_return(double('response', :code => 200))
@@ -76,6 +92,14 @@ describe Localeapp::ApiCaller, "#call(object)" do
     it "doesn't set the HTTPClient ca_file if ssl_ca_file is nil" do
       Localeapp.configuration.ssl_ca_file = nil
       RestClient::Request.should_receive(:execute).with(hash_not_including(:ca_file => nil)).and_return(double('response', :code => 200))
+      @api_caller.call(self)
+    end
+  end
+
+  context "Timeout" do
+    it "sets the timeout to the configured timeout" do
+      Localeapp.configuration.timeout = 120
+      RestClient::Request.should_receive(:execute).with(hash_including(:timeout => 120)).and_return(double('response', :code => 200))
       @api_caller.call(self)
     end
   end