+++ spec/localeapp/cli/push_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 describe Localeapp::CLI::Push, "#execute(path)" do
   let(:output) { StringIO.new }
-  let(:pusher) { Localeapp::CLI::Push.new(output) }
+  let(:pusher) { Localeapp::CLI::Push.new(:output => output) }
 
   context "path is a directory" do
     it "calls push_file on each yaml file in the directory" do
@@ -33,7 +33,7 @@ end
 
 describe Localeapp::CLI::Push, "#push_file(file_path)" do
   let(:output) { StringIO.new }
-  let(:pusher) { Localeapp::CLI::Push.new(output) }
+  let(:pusher) { Localeapp::CLI::Push.new(:output => output) }
 
   it "creates a new file object and makes the api call to the translations endpoint" do
     with_configuration do