+++ spec/localeapp/routes_spec.rb
@@ -82,6 +82,74 @@ describe Localeapp::Routes do
     end
   end
 
+  describe "#remove_endpoint(options = {})" do
+    it "returns :delete and the remove url for the options" do
+      with_configuration(@config) do
+        options = { :key => 'foo.bar' }
+        @routes.should_receive(:remove_url).with(options).and_return('url')
+        @routes.remove_endpoint(options).should == [:delete, 'url']
+      end
+    end
+  end
+
+  describe "#remove_url(options = {})" do
+    it "it extends the project_url and includes the escaped key name" do
+      with_configuration(@config) do
+        @routes.remove_url(:key => 'test.key').should == "https://test.host/v1/projects/API_KEY/translations/test%2Ekey"
+      end
+    end
+  end
+
+  describe "#rename_endpoint(options = {})" do
+    it "returns :post and the rename url for the options" do
+      with_configuration(@config) do
+        options = { :current_name => 'foo.bar' }
+        @routes.should_receive(:rename_url).with(options).and_return('url')
+        @routes.rename_endpoint(options).should == [:post, 'url']
+      end
+    end
+  end
+
+  describe "#rename_url(options = {})" do
+    it "it extends the project_url and includes the escaped key name" do
+      with_configuration(@config) do
+        @routes.rename_url(:current_name => 'test.key').should == "https://test.host/v1/projects/API_KEY/translations/test%2Ekey/rename"
+      end
+    end
+  end
+
+  describe "#export_url" do
+    it "it extends the project_url and defaults to yml" do
+      with_configuration(@config) do
+        @routes.export_url.should == "https://test.host/v1/projects/API_KEY/translations/all.yml"
+      end
+    end
+
+    it "adds query parameters on to the url" do
+      with_configuration(@config) do
+        url = @routes.export_url(:query => {:updated_at => '2011-04-19', :foo => :bar})
+        url.should match(/\?.*updated_at=2011-04-19/)
+        url.should match(/\?.*foo=bar/)
+      end
+    end
+
+    it "can be changed to another content type" do
+      with_configuration(@config) do
+        @routes.export_url(:format => :json).should == 'https://test.host/v1/projects/API_KEY/translations/all.json'
+      end
+    end
+  end
+
+  describe "#export_endpoint(options = {})" do
+    it "returns :get and the export url for the options" do
+      with_configuration(@config) do
+        options = { :foo => :bar }
+        @routes.should_receive(:export_url).with(options).and_return('url')
+        @routes.export_endpoint(options).should == [:get, 'url']
+      end
+    end
+  end
+
   describe "#missing_translations_endpoint(options = {})" do
     it "returns :post and the missing_translations url for the options" do
       with_configuration(@config) do